Use classification: branch to avoid tricky executive / legislature distinctions
This is a temporary workaround; the city councils are being used as
executives *and* legislative houses at the moment for city councils,
which is causing problems because they appear with the same ID but two
different classifications downstream when trying to merge all the
Canadian data together. This isn't the right solution to that problem,
but it'll do for the moment.